requirements.txt
boto3==1.24.71
botocore==1.27.71
psycopg2-binary==2.9.3
